python_advanced_diploma/backend/requirements/base.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0